python-embed/requirements-real.txt
# Optional dependencies for EMBED_MODEL=dog-nose-identification2
# Keep these out of the default mock image to preserve lightweight CI.

numpy==1.26.4
Pillow==10.3.0
timm==1.0.3
torch==2.3.1
torchvision==0.18.1
